Add extra-dirty KTX2 suport for Linux
This is not a proper port, and is just a list of changes that were
necessary to build it on my machine. And then add KTX2 support.

Requires building KTX library manually after fetch_dependencies.py
checkout, which is not great.

Lots of changes need to be cleaned too. It's a deliberate mess right
now.